Change responsibility from the repo to the service
A topic kafka is phiysically on a cluster. Then we apply the Access
Controll Entry with the namespace to manage rigths.

So to correspond to the low level system, I changed the interface of repository to
remove reference to the namespace

Now the service managed the access of the topic via the Namespace